Hardcode devices + remove rootDir
