--- cli/command/init/README.md
@@ -1,19 +1,0 @@
-# saucectl init
-
-# Frameworks
-
-- Cypress
-- TestCafe
-- Playwright
-- Puppeteer
-- Espresso
-- XCUITest
-
-# Workflow
-
-- Framework Selection
-  - Cypress
-    Version Selection (3 majors)
-      - 7.6.0
-      - 6.5.0
-      - 5.6.0
