+++ cli/command/init/lists.go
@@ -1,0 +1,8 @@
+package init
+
+var androidDevicesPatterns = []string{
+	"Amazon Kindle Fire .*", "Google Pixel .*", "HTC .*", "Huawei .*",
+	"LG .*", "Motorola .*", "OnePlus .*", "Samsung .*", "Sony .*",
+}
+
+var iOSDevicesPatterns = []string{"iPad .*", "iPhone .*"}
